requirements.txt
pandas
numpy
scikit-learn
pyyaml
joblib
uvicorn==0.27.1
